Separate messaging coroutinescope class
