+++ src/main/kotlin/fi/metatavu/vp/vehiclemanagement/rest/AbstractApi.kt
@@ -1,6 +1,6 @@
 package fi.metatavu.vp.vehiclemanagement.rest
 
-import fi.metatavu.vp.WithCoroutineScope
+import fi.metatavu.vp.vehiclemanagement.WithCoroutineScope
 import jakarta.inject.Inject
 import jakarta.ws.rs.core.Context
 import jakarta.ws.rs.core.HttpHeaders
